Expose associate_public_ip_address in launch configuration
Sometimes it might be useful to have your ASG internet accessible, for
example if you are not using a load balancer and need the to expose the
services on them directly, or if you don't have a bastion. This update
exposes the associate_public_ip_address option in the launch
configuration.

Note that this does not mean that your hosts are internet accessible
other than via ICMP - you still need to open up the ports via security
group rules.