Add extra_instance_tags
This update adds extra_instance_tags, which allows the user to supply
a list of hashes with the same syntax as the tag parameter for the
aws_autoscaling_group resource. This allows one to supply extra tags
that will get added to the ASG, and propagate to its instances so long
as propagate_on_launch is set, which is more than likely what the user
wants. A note has been added to the docs to ensure that is noted.